基于文件路径的访问 有BUG

+++ code/kernel/src/boards/qemu.rs
@@ -1,7 +1,7 @@
 #[cfg(not(feature = "board_hifive"))]
-pub const CLOCK_FREQ: usize = 12500000;
+pub const CLOCK_FREQ: u128 = 12500000;
 #[cfg(feature = "board_hifive")]
-pub const CLOCK_FREQ: usize = 1000000;
+pub const CLOCK_FREQ: u128 = 1000000;
 
 // pub const MMIO: &[(usize, usize)] = &[(0x10001000, 0x1000)];
